renamed exmeption cert to be more appropriate
